ai-orchestrator-service/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
google-generativeai==0.8.3
pydantic==2.10.0
httpx==0.28.0
zeep==4.3.1
gql[httpx]==3.5.0
grpcio==1.68.0
grpcio-tools==1.68.0
protobuf==5.28.0
python-dotenv==1.0.0
